refactor: move gRPC converters to separate file
